requirements/base.txt
Django==3.0.2
djangorestframework==3.11.0
psycopg2==2.8.4
djoser==2.0.3
boto3==1.11.6
django-fernet-fields==0.6.0
